move checkWatchListDataConsistencyIfRequested back to client-go/tools/cache
